requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
pyodbc==5.0.1
alembic==1.12.1
pandas==2.1.4
rapidfuzz==3.5.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
nltk==3.8.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
